requirements.txt
asgiref==3.7.2
autopep8==2.0.4
black==23.9.1
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==0.7
Django==4.2.4
django-crispy-forms==2.0
Faker==19.4.0
isort==5.12.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
pep8==1.7.1
Pillow==10.0.0
platformdirs==3.10.0
pycodestyle==2.11.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
